Adjust `net.core.rmem_max` and `net.core.wmem_max` on the host to make Quic-Go happy
Without this, if `h3://` upstream DNS servers are defined, AdGuard Home (Quic-Go within it)
will report:

> failed to sufficiently increase receive buffer size (was: 208 kiB, wanted: 7168 kiB, got: 416 kiB). See https://github.com/quic-go/quic-go/wiki/UDP-Buffer-Sizes for details.

Ref: https://github.com/quic-go/quic-go/wiki/UDP-Buffer-Sizes

This is done on the host, not inside the container (via `docker run --sysctl`),
because the latter doesn't seem to work.

We only increase the values of `net.core.rmem_max` and `net.core.wmem_max` to the specified
minimium values. If the host uses larger values for these sysctl parameters,
we retain what the host uses and don't decrease it.